pyproject.toml
[tool.poetry]
name = "ttsclient"
version = "1.0.23"
description = ""
authors = ["wok <test@test.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10,<3.11"
fire = "0.6.0"
pythonnet = "3.0.3"
numpy = "1.26.4"
pytorch-lightning = "2.4.0"
matplotlib = "3.9.1.post1"
requests = "2.32.3"
einops = "0.8.0"
transformers = "^4.48.0"
cn2an = "0.5.22"
pypinyin = "0.52.0"
jieba-fast = "0.53"
librosa = "0.10.2.post1"
scipy = "1.14.0"
fastapi = "0.112.0"
portpicker = "1.6.0"
uvicorn = "0.30.5"
python-socketio = "5.11.3"
python-multipart = "0.0.9"
pyopenssl = "24.2.1"
aioquic = "1.2.0"
fastapi-cors = "0.0.6"
ngrok = "1.4.0"
uuid = "1.30"
wordsegment = "1.3.1"
g2p-en = "2.1.0"
jamo = "0.4.1"
ko-pron = "1.3"
g2pk2 = "0.0.3"
g2pw = "0.1.1"
opencc = "1.1.9"
pyjyutping = "1.0.0"
simple-performance-timer = "0.1.3"
onnx = "1.16.2"
torchmetrics = "1.4.2"
onnxsim = "^0.4.36"
x-transformers = "^2.1.2"
torchaudio = {version = "^2.6.0", source = "torch_cu128"}
jieba = "^0.42.1"
split-lang = "^2.1.0"
nltk = "3.8.1"
faster-whisper = "^1.1.1"
onnxruntime = "1.20.1"
peft = "^0.14.0"
fast-langdetect = "^0.3.1"
tojyutping = "^3.2.0"
onnxruntime-gpu = "1.20.1"
torch = {version = "^2.6.0", source = "torch_cu128"}
pyopenjtalk-plus = "^0.4.1.post7"


[tool.poetry.group.dev.dependencies]
pip-licenses = "^4.4.0"
toml = "0.10.2"
types-requests = "^2.31.0.20240406"
pytest = "^8.1.1"
types-pyopenssl = "^24.0.0.20240417"
types-pycurl = "^7.45.3.20240421"
types-toml = "^0.10.8.20240310"
types-tqdm = "^4.66.0.20240417"

[[tool.poetry.source]]
name = "torch_cu128"
url = "https://download.pytorch.org/whl/cu128"
priority = "explicit"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
[tool.poetry.scripts]
main = "ttsclient.main:main"
generate_version_file = "scripts_.version_manager:main"
generate_license_file = "scripts_.license_manager:generate"
